completions: support completion of jobids with ascii f
Problem: The bash completions suggest jobids in f58 encoding, but users
may want to complete jobs using a plain f since it is easier to type.

For commands that can complete with jobids, if the current string
starts with an ascii f, then force the suggested jobid list to be in
f58plain encoding.