shell/stage-in: accept old stage-in.tags option
Problem: stage-in.tags (the old option) is now silently ignored.
This might cause jobs that are pending in the queue when Flux
is updated to misbehave.

If stage-in.tags is set, print a warning on the leader shell, but
accept it as an alias for the new option, stage-in.names.

In addition, add "!" to the option parser so any unknown options
trigger a hard failure and are not simply ignored.