testsuite: test legacy filemap and stage-in options
Problem: there are no tests for flux-filemap or for
stage-in.tags.

Add a sharness test for flux-filemap.  The tests are not mixed
in with those of flux-archive so they can be more easily removed
when we decide to drop backwards compatibility.

Add a test for to t2617-job-shell-stage-in.t for the tags option.
Also, check that unknown options are detected rather than ignored.