job-manager: omit zombies from job-manager.list
Problem: zombies (inactive, waitable jobs) are unconditionally
listed by the job-manager.list RPC, and this list is expected
to be empty by various tests, which break when waitable semantics
change.

Omit zombies from the RPC response.

Update t2208-job-manager-wait.t which was using that output, but
not to much benefit.