job-exec: make job exception optional in lost_shell()
Problem: When a noncritical job shell is lost, the job-exec module
always raises a nonfatal job exception and sends an RPC to the
leader shell exception service. However, it may be useful to only
notify the job shell and not clutter the eventlog with an exception
in some circumstances.

Add a raise_exception parameter to lost_shell(). A job exception will
not be raised if this parameter is false.