shell: output: track active shell ranks for proper refcounting
Problem: If a shell is lost after its output plugin sends the eof
message to the leader, then the leader could double decrement its
internal reference count and exit before other shells have completed.

Track active job shells in an idset and only decrement the refcount
if a lost job shell or EOF was from a shell that was active.

Fixes #2492