flux-job: add shorthand paths to eventlog, wait-event
Problem: The eventlog and wait-event subcommands of flux-job(1) take
a `-p, --path=PATH` option to specify alternate eventlogs on which
to operate, but these options require the full KVS path in the job
namespace which can be annoyingly lengthy, e.g. guest.exec.eventlog.
Since these eventlogs are standard and may be commonly used with the
command, it is not user-friendly to require full paths in this case.

Support short names for well-known standard eventlog paths such as
'exec' for 'guest.exec.eventlog', 'output' for 'guest.output' and
'input' for 'guest.input'.