flux-job: split MPIR interface into separate source file
Problem: The MPIR support code for flux-job is within the same source
file as the code for the attach command. As the MPIR support for
flux-job is expanded, it will be cleaner to keep the majority of the
MPIR support contained in a separate source file.

Split the MPIR symbols and proctable generation routines into their
own source file in src/cmd/job/mpir.c.