testsuite: switch mpir test program to optparse
Problem: The t/shell/mpir.c test program uses free arguments
to pass the leader shell rank and service name to the test, but this
is not flexible enough for future use of this program.

Add `-r, --leader-rank=N` and `-s, --service=NAME` options to the
test program, and drop the use of free arguments.

Update the affected sharness test.