testsuite: fix random-generation of kvs archive files in t0021
Problem: the randbytes() function in t0021-archive-cmd.t uses
/dev/random to create files of size 128 and 2048 bytes, respectively.
However, /dev/random can be truncated at <128 bytes and, when it is,
the file created by dd will be as well. This causes a size mismatch
in the kvs, which causes tests 13, 14, and 18 to fail. This first
showed up in GitLab CI on TOSS 4.

We can instead utilize the /dev/urandom file, which is not truncated,
and thus guaranteed to create files of the proper sizes. Also, some
cleanup of t0021-archive-cmd.t is baked into this PR.

When merged, closes issue #5750.