testsuite: test correctness of shell rankinfo object
Problem: There is no test that ensures the job shell rankinfo object
as returned by flux_shell_get_rank_info(3) contains the correct values
for broker rank and number of tasks for multiple job shell ranks.

Leverage the rc scripts and add a test to t2603-job-shell-initrc.t
for this purpose. Increase the number of brokers in the test to allow
for testing with mulitple job shells in a single job.