shell: mpir: compress repeating ranges in rangelist
Problem: A common case for MPIR proctable entries for large jobs is
to have many repeating range entries in a rangelist, e.g.

 [123,-7],[1,-7],[1,-7],[1,-7]

This could be easily further compressed.

Support range repeats in the JSON representation of a rangelist. Add
an optional third member of the range array which, when present,
indicates that the range itself is duplicated N times. The above
example then becomes

 [123,-7],[1,-7,2]