flux-job: add hostpids subcommand
Problem: Third party tools may support an option to pass a set of
hostname, PID pairs for operating on tasks of a parallel job, but
there is currently no convenient way to get this information for
a job.

The job shell exports this data via the proctable RPC used to capture
the MPIR_proctable in flux-job attach. Add a new 'hostpids' subcommand
to flux-job(1) which uses the MPIR support to capture the proctable
from the job shell and prints a comma-separated set of host:pid pairs
to stdout.

The command works for pending as well as running jobs so that a tool
can be attached to a running job with something like

 $ tool --hostpids=$(flux job hostpids $(flux submit ...))

The command offers options to change the delimiter select a subset
of task ranks, and apply a timeout.