admin-guide: clarify queue property name requirements
Problem: There is a statement in the admin guide that it is good
practice to use property names that match the corresponding queue
name, but no reason for this practice is given.

Explain that queue properties do not need to match the queue name,
but note that it is good practice because it allows flux-resource
list to suppress these duplicate names in is output.