shell: output: support append to output file
Problem: There is no way to tell the shell output plugin to append
to a file instead of truncating on open(2).

Add support for a new output plugin option: `output.mode`. The default
is "truncate", which preserves current behavior. If output.mode is
set to "append", then `O_APPEND` will replace `O_TRUNC` in the flags
passed to open(2).