testsuite: fix lingering processes after `make check`
Problem: There are often a couple orphaned Flux instances that remain
after a run of `make -j N check`. The instances appear to be blocked
in rc3 with a sleep job in CLEANUP which can never exit because the
scheduler module has already been unloaded. The root cause of this
situation is unknown, but probably the affected tests can be improved
to reduce the probability of the issue.

The affected tests include t2800-jobs-recursive.t and t2802-uri-cmd.t,
each of which start a nested instance of one or more levels with
sleep jobs.

Attempt to terminate the affected batch jobs cleanly rather than
leaving it to the test instance rc3 script. This should reduce or
eliminate stray processes after running make check or the individual
test.