job-manager: perilog: support list of log patterns to ignore
Problem: The job-manager perilog plugin logs all stdout and stderr
lines from the prolog and epilog to the broker log buffer. On systems
with lots of nodes, even a line of useless output per node can fill
the logs with noise, but we don't want to suppress all log output.

Support a new configuration value for the perilog plugin

   job-manager.perilog.log-ignore = [ LIST ]

to supply a LIST of regex(7) patterns to ignore in prolog and epilog
lines of output.

The patterns support POSIX extended regex syntax.

An empty line of output is always ignored.