job-exec: allow formatted message to be passed to lost_shell()
Problem: The lost_shell() function that raises a job exception
and notifies the leader of lost shells constructs a message only
for the job exception, and that message is specific to a node failure
and doesn't capture other reasons a shell might be lost. It would be
better to be more flexible.

Support passing the message as printf-style arguments instead of
only taking a hostname in lost_shell(). If a job exception is being
raised, do not also pass the message in the shell exception RPC
to avoid duplication.