job-manager: include R in sched.free request
Problem: R has to be looked up from the KVS in the sched.free
request handler, but now that the job manager caches R, this
is an unnecessary extra step.

Add R to the sched.free request payload.

Note that the `R.scheduling` key is not included.  The current design of
Fluxion in which `R.scheduling` may contain a voluminous JGF object made
caching this part of R impractical.

Change libschedutil so that
- the sched.free message handler never looks up R in the kvs
- the free callback always sets its `R` argument to NULL
- the SCHEDUTIL_FREE_NOLOOKUP flag is a no-op

Update sched-simple's free callback to unpack R from the message
instead of decoding the `R` arugment.

Note that Fluxion sets SCHEDUTIL_FREE_NOLOOKUP so it already expects
the free callback's R argument to be NULL.  Although this change increases
the size of sched.free payloads with data that Fluxion currently does not
use, the ranks in R will be required by Fluxion in the future to identify
resource subsets for partial release (flux-framework/flux-sched#1151).

This change should be accompanied by an update to RFC 27.

Update sched-simple unit test.

Fixes #5775