job-exec: do not duplicate lost shell message
Problem: If a message is set when lost_shell() is called, it will be
displayed twice: once when the exception note is printed, and another
when the shell displays the message included with the exception RPC
payload.

If an exception is raised on the job, clear the message so the shell
won't bother displaying it.