docker: add script help message for macOS
Problem: recently, several core developers have noted that the
docker-run-checks.sh script borks when command-line options that specify
values do so with an equals sign rather than a space on macOS.

Add a help message to show up on macOS reminding devs to strip all '='
and replace them with spaces in their command-line arguments if BSD
getopt is used. Also, give users the option to set an environment
variable to tell the script on macOS that it should assume GNU getopt
is being used ('FORCE_GNU_GETOPT=1').