job-manager: drop sched.free response requirement
Problem: the sched.free RPC requires a response but the response
is unnecessary for correct job management.

Moreover, the coupling between the RPC response and the job state
machine makes some anticipated changes harder to design.

Make sched.free a "fire and forget" RPC.  Simply post the free event
immediately instead of deferring it until the response is received.

Update libschedutil to make schedutil_free_respond() a no-op.
Update sched-simple to not directly respond when there is a free error.

This change should be accompanied by an update to RFC 27.