modules/content: drop incorrect assertion
Problem: sometimes brokers hit an assertion when the
content module is unloaded during shutdown.

The assertion is that there are no pending load/store operations
waiting on a cache entry when it is destroyed.  This is certainly
possible, depending on the behavior of users of the load/store
RPCs.  In the case of unload in rc3, dependent modules would have
already been unloaded, so it seems safe to simply destroy any
pending RPCs without responding to them.

Destroy them, if present.

Fixes #5706