resource: call status disconnect handler
Problem: when a client disconnects with a resource.status RPC
pending, the RPC is not aborted.

Change the resource module disconnect handler to allow non-owners
to send the message (and verify that the acquire handler is using
proper authentication).  Then call status_disconnect() from the
handler.

Since the disconnect handler also calls acquire_disconect() and
that only works on rank 0, add a check so it doesn't segfault.