libsubprocess: fix possible use-after-free in remote process kill
Problem: When flux_subprocess_kill(3) is called and the subprocess
is not yet running, an empty future is returned that will later be
fulfilled when a subsequent kill RPC is initiated when the process
does start running. However, if the future returned to the caller
is destroyed between when the actual RPC is initiated and the
sigpending_cb() is called, then the orignial future (stored in the
subprocess object) will be used after it is freed and a segfault or
memory corruption can occur.

If the future returned from flux_subprocess_kill(3) is destroyed, then
any delayed RPC future should also be destroyed, thus canceling the
sigpending_cb() callback. Ensure this is the case by tying destruction
of the RPC future to the flux_subprocess_kill(3) future by setting
a destructor in the latter's aux item list.

Since the RPC future is now "owned" by the future returned from
flux_subprocess_kill(3), no longer destroy it in sigpending_cb().

Take a reference on the returned future in add_pending_signal()
and drop it in fwd_pending_signal() to handle the case where the
caller destroys the future between these two calls (e.g. as in
fire-and-forget). This ensures that installing a destructor for the
RPC future created in fwd_pending_signal() is successful. Once the
reference on the original future is dropped, the RPC future may be
destroyed immediately if this was the last reference to the original
future.

Fixes #5782