job-exec: reduce useless log noise
Problem: When terminating a job in the midst of starting, or when a
fatal job exception is raised on a job that has already been canceled,
the job-exec module may generate lots of log messages like:

 exec_kill: any (rank 4294967295): No such file or directory

In this case it is likely the subprocess was unable to be signaled,
therefore a broker rank is not associated with the child future for
which the erorr is being printed.

It isn't valuable to print an error in this case, so just skip the
flux_log() when rank == FLUX_NODEID_ANY.

Similarly, it probably isn't useful to log an error if the errno is
ENOENT, since this just means the remote process is no longer running.