kvs: remove excessive logging
Problem: When a namespace is marked for removal, it is normal to
cleanup lingering/left over transactions that are not yet complete.
On some tests, this appears to lead to many unnecessary log messages.

Remove log messages related to cleanup of transactions on removed
namespaces.  They are unnecessary and excessive.  Those transactions
will receive a ENOTSUP error and can report that the namespace has
been removed.

Fixes #5789