job-exec: fix potential namespace leak
Problem: The job->has_namespace flag is only set after the response
for the KVS namespace creation RPC is received. This can cause a rare
namespace leak if a job exception occurs after the request is sent,
but before the response is received. The cleanup code may not see
that the has_namespace flag is set and the namespace is never deleted.

Set job->has_namespace just after sending the namespace create RPC, so
that namespace cleanup is forced if there is any chance the namespace
will exist when the job is complete.

Also, to avoid skipping namespace_delete() when some other previous
function fails (such as the final eventlog write or the copy of
the namespace contents), call flux_future_or_then(3) as well as
flux_future_and_then(3) in the chain of namespace cleanup futures.

Fixes #5790