ci: update black requirement from 22.8.0 to 24.3.0
Problem: Github dependabot notified this project of a security
issue with black 22.8.0 and recommends updating to 24.3.0.

Update black requirement to 24.3.0 and commit all formatting changes
as a result.

Also make the following other required changes to get this version
of black to work:

Ensure the python linting action runs on ubuntu-latest in github,
otherwise this version of black is not found by pip.

Update pyproject.toml to remove tool.black.profile, which the newer
black doesn't understand or support.

Update pyproject.toml to remove tool.black.exclude, which wasn't
being used under pre-commit anyway. Replace with force-exclude,
which applies the exclusion pattern even if the targets are passed
as a list of files as is done with pre-commit.