testsuite: fix test racing with exit-timeout
Problem: The t4583-free-range-test.sh test will kill a broker
to emulate a node failure.  On a slow system or a parallel run
of the testsuite, this can lead the test to race against the
exit-timeout, leading to an unexpected test failure.

Solution: set exit-timeout=none for the job allocated in the
test.

Fixes #5809