testsuite: correct status within test comments
Problem: A comment in test t4583-free-range-test.sh suggests
a broker overlay status should be viewed as "offline".  However that
shouldn't be the case, since the broker previously joined the
instance.  An "offline" status indicates the broker never joined
the instance.

Update comment to say that the broker should be "lost".