job-exec: always notify shell rank 0 of lost shells
Problem: When a critical job shell is lost, the job execution service
raises a fatal job exception, but does not also notify the leader
shell that another shell has been lost. This can lead to the leader
shell waiting unnecessarily for EOF from the lost shell, which delays
job exit.

Always notify the leader (shell rank 0) of lost shells. Tweak the
call signature of lost_shell() to take a 'critical' parameter that
indicates if the lost shell was critical or not to make the intent
more obvious. Only raise a non-fatal (informational) job exception
for non-critical shells, and raise a fatal job exception for critical
shells.

Fixes #5811