job-exec: raise fatal error if leader job shell exits with signal
Problem: The job execution system notifies the leader job shell when
other job shells are lost due to an abnormal exit to avoid a hang at
job termination where the leader shell is waiting for EOF or other
events from missing job shells. However, if the rank 0 shell is lost
due to an abnormal exit like a signal, then the job could also hang,
since the other job shells may not realize the leader shell is gone.

Raise a fatal job exception and start job termination if the rank 0
job shell exits due to a signal to avoid this potential hang.