flux-resource: improve performance of ResourceSetExtra initializer
Problem: The ResourceSetExtra class is supposed to be just a wrapper
around a ResourceSet class that adds the convenience `propertiesx`
and `queue` properties. However, using this class is slow because
it reinvokes the underlying ResourceSet initializer, which ends up
creating a unnecessary copy of the resource set from scratch.

Make the class a wrapper by just stashing the original ResourceSet
object and forwarding unknown getattrs to the wrapped object. This
avoids the wasted time recreating the coped resource set.