flux-resource: increase efficiency of resource_uniq_lines()
Problem: The `flux resource list` command spends the majority of its
time in resource_uniq_lines() on large clusters because it iterates
over every rank in each state resource set in order to create the
set of unique lines.

Instead split the resource into smaller subsets based on all
combinations of properties contained within that set. For the purposes
of the current incarnation of `flux resource list` this should be
the minimum number of distinct resource sets required to generate
possibly unique lines of output.