librlist: improve performance of rlist construction
Problem: Profiling shows that an inordinate amount of time is spent
in util_idset_add_check() when building an rlist object from an
R object. This appears to be beacuse both the idset calls in this
function; idset_has_intersection() and idset_union(); always traverse
all of the idset `b` entries. In the case of building a new rlist,
the idset `a` is empty because we are adding new ids to a new rnode,
so this traversal is completely unnecessary.

In the case that idset `a` is empty, return a copy of `b` and skip
calling the other idset calls completely.