python: use resource.sched-status
Problem: 'flux resource list' can take a long time when the scheduler
is busy.

Use resource.sched-status instead of sched.resource-status to
request resource information.  Force the RPC to go to rank 0 since,
unlike the scheduler, resource is loaded on all ranks and would
fail the RPC if first sent to the local broker.