flux-top: support FLUX_RESOURCE_LIST_RPC environment variable
Problem: There is no easy way to test flux-top(1) fallback to the
sched.resource-status RPC topic string.

Use the FLUX_RESOURCE_LIST_RPC environment variable established for
use in the Python bindings to override the default topic string that
flux-top(1) uses to get the scheduler view of resources. This can be
set to sched.resource-status to avoid the fallback code completely
when connecting to a known older instance, or can be set to an invalid
topic string for testing the fallback code.