job-ingest: reserve the top 16 FLUID generators
Problem: job-ingest is loaded on up to 2^16 ranks, using up all
valid FLUID generator IDs, which prevents a specialized range from
being used later.

A specialized range might include IDs allocated directly by the
job manager for high throughput such as discussed in #3556.

Reserve the top 16 generator IDs.  As before, the job-ingest module
simply exits with a return code of 0 on the ranks that are out of range
and job ingest is handled upstream.

Since this logic was incorrect earlier and we missed it, add a
command line option: max-fluid-generator-id=N to be used in testing.