job-manager: send complete journal to consumers
Problem: consumers of the journal have to scan the KVS, a costly
operation, to obtain the full set of job IDs and their job state,
then reconcile possibly overlapping events received in the journal.
This protocol can be simplified now that the job manager stores
inactive jobs and their eventlogs.

Change the journal response message payload from:
  {"events":[
    {"id":I "eventlog_seq":i "entry":o}
    {"id":I "eventlog_seq":i "entry":o}
    {"id":I "eventlog_seq":i "entry":o}
    ...
  }
to
  {"id":I "events":[
    o
    o
    o
    ...
  ]}
where 'o' is a RFC 18 eventlog entry.

Upon receipt of a request by a journal consumer, the job manager sends a
response message for each active job.  If the "full" flag is set in the
request, inactive jobs are sent also.  Once the events for existing jobs
have been sent, the job manager sends a sentinel response:
  {"id":FLUX_JOBID_ANY, "events":[]}
The purpose of the sentinel is to allow a journal consumer to delimit
the stream of historical events from new events, in case it wishes to
ingest all historical events as part of its initialization.

Following the sentinel, the job manager sends a journal response message for
each new event.

The per-job event sequence number and associated logic has been dropped
since now the potential for overlapping events between KVS and journal
has been eliminated.

The 'job-manager.journal-size-limit' config key is dropped since journal
streams are now generated directly than pushed through a shared list.