testsuite: test flux-top(1) fallback to sched.resource-status
Problem: There is no test that ensures flux-top(1) falls back to the
old sched.resource-status RPC when the primary RPC returns ENOSYS.

Use FLUX_RESOURCE_LIST_RPC to set a invalid topic for the initial RPC,
forcing the fallback code to be activated. Ensure resources are still
displayed in flux-top(1).