job-manager: improve replay warning message
Problem: A warning in job-manager/restart.c prints the job state
after event_job_action() is complete, but it is more useful to know
the state of the job before event_job_action().

Save the original job state and print both the state before and after
event_job_action() since this information will be helpful when this
error occurs.