job-list: defer requests until after init
Problem: job-list requests should not be processed until the
module is fully initialized.

While jsctx->initialized is false, requests are placed on a queue.
When jsctx->initialized is set true after the initial journal backlog
processing, messages on the queue are requeued in the flux_t handle
for reactive processing.