flux-dump: call sd_notify()
Problem: if an instance that is configured to dump is shutdown
via 'systemctl stop flux', it could run out the shutdown timer and
be killed, interrupting the dump.

Have flux-dump check for the broker attribute 'broker.sd-notify'.
If it is set to something other than zero, then use sd_notify(3) to
send systemd EXTEND_TIMEOUT_USEC messages as the dump progresses.
In addition, send progress updates that are displayed under "Status:"
in 'systemctl status flux'.