python: return empty string on epoch time
Problem: The conversion flag "D" returns an empty string when the
value of a timestamp is not present.  In some cases, timestamps are
initialized to 0.0 if the timestamp is not available.  This leads
to the "D" conversion flag unexpectedly returning the epoch time
1969-12-31T16:00:00 instead of the empty string.

Handle epoch time as an additional special case for the D conversion
flag.  When the timestamp is 0, have the D conversion flag return
an empty string.

Fixes #5827