flux-shutdown: add --skip-gc option
Problem: when garbage collection is enabled by default in systemd,
it may be handy to have a way to avoid it, for example if the system
has been garbage collected recently, but it is still expected to
take a long time.

Add a flux shutdown --skip-gc option to avoid garbage collection.