job-manager: include R, jobspec in journal stream
Problem: job-list, the primary consumer of the job manager
journal, has to look up R and jobspec in the KVS, but the job
manager already has redacted versions of these in memory,
which should be suitable.

When generating the initial backlog, send them along if set in
the job structure.

When processing live events, send jobspec with the validate
event and R with the alloc event.