job-list: improve journal error handling
Problem: ENODATA is not treated specially (as per RFC 6) when
the journal RPC gets an error response.

Detect that case, log it, and stop the reactor without error.
Also, don't decode the eventlog initally, as it is not used yet.
Let any (unlikely) protocol errors be handled when the message is
processed.