shell: support hostfile taskmap scheme
Problem: There is currently no simple way to map tasks to specific
hosts using a list of hostnames.

Add a hostfile taskmap plugin to the job shell. This taskmap scheme
assigns tasks to hosts in order as they appear in a file provided by
the scheme value. The hostfile is read as a list of newline separated
hostnames or RFC 29 Hostlist strings. If there are not as many hosts
as tasks in the hostfile, then the list is reused, and if more hosts
than tasks appear in the file, then the list is truncated.

Like other taskmap schemes, it is an error if the same task count is
not assigned to each host as in the default task map.

Fixes #5794