job-manager: add plugin to assist posting manual job events
Problem: Sometimes a job becomes stuck in a given state, requiring
intervention by forcing an event to be posted to the job manager.
Currently, this requires building a custom jobtap plugin, which is
annoying and not sysadmin friendly.

Add a builtin jobtap plugin that will simply post a custom job
event defined in the RPC.