cmd: add flux-post-job-event.py
Problem: There's no convenient way to post job events manually when
it's necessary.

Add a new `flux post-job-event` command which sends an RPC to the
job-manager.post-event.post service when called, e.g.:

  flux post-job-event JOBID NAME [KEY=VAL, KEY=VAL..]

This command is meant to be a temporary solution to clean up cases
where a job gets stuck in CLENAUP becuase an epilog-start event was
posted without a corresponding epilog-finish event.