job-sql: add job query service with SQL capability
Problem: Flux doesn't have a raw SQL interface to job data that
can utilize the sqlite JSON1 extensions.

Add a service that consumes the job manager journal and populates
an in-memory sqlite database with all jobs (active and inactive jobs).
The schema simply stores the jobid, eventlog, jobspec, and R.
The last three are kept in JSON format so sqlite JSON1 extensions
can be used to construct queries.

https://www.sqlite.org/json1.html