broker: run shutdown script not cleanup commands
Problem: the cleanup commands pushed into broker memory in rc1
are not easily maintained or extended.

Create a new "shutdown" script that lives next to rc1 and rc3.
Run this script instead of the "cleanup" commands when the broker
CLEANUP state is entered.  Upon completion of the shutdown script,
the broker transitions to the SHUTDOWN state.