resource: add drainset class for drain response handling
Problem: The generation of the drain response in drain_get_info()
is very slow when there are a moderate amount of drained ranks.

Add a new 'drainset' class which is an abstraction for a set of drained
targets, each with a timestamp and reason field. This can be used
to replace the slow, in-place generation of the drain info JSON object
in resource/drain.c.