broker: allow cleanup-push to continue to work
Problem: some framework projects are using "test personalities"
that call flux admin cleanup-push.

Accept "cleanup" as an alias for "shutdown" to allow those projects
to continue to work while transitioning to the shutdown script.

Note: although these commands are no longer interpreted by the shell,
none of those projects appear to be relying on that.