testsuite: fix invalid config tests in t2312
Problem: The invalid config tests in t2312-resource-exclude.t
fail because there is no tty, not due to the invalid config. This
is because the instances are run without argument, so are trying to
launch an interactive shell. The tests thus fail with:

  lt-flux-broker: stdin is not a tty - can't run interactive shell
  ok 7 - config with bad exclude idset fails

  lt-flux-broker: stdin is not a tty - can't run interactive shell
  ok 8 - config with out of range exclude idset fails

Add `true` as the argument to flux-start so that the instance would
succeed if the invalid config was ignored.