python: improve resource.status with many drained ranks
Problem: The ResourceStatus class performs a couple unnecessary idset
set operations for each drain entry in the resource.status response
object. When there are many entries, these set operations perform a
lot of unnecesary work and slow down the _recalcultate() method.

Build the set of drained ranks in the loop (which should be fast),
and delay splitting the set into drained and draining after the loop.