python: fix performance of ResourceStatus.get_drain_info()
Problem: The ResourceStatus get_drain_info() method is very slow
when there are many drained ranks because it uses a linear search.
Since it will be called at least once per rank, it causes an
exponential performance hit based on the number of drained ranks.

Add a mapping from rank to DrainInfo object in the ResourceStatus
class and use this instead of a search in get_drain_info().

Fixes #5859