flux-resource: fix performance issue generting queue output
Problem: In commit 42438952bc0933b84b4bdf16786f643575853a5a a test
for an empty resource set was added to the queue property of the
ResourceSetExtra class to avoid generating a queue for an empty
resource set. However, the test for empty uses self.rlist, which is an
expensive call. Also, the test is done for each queue in the config,
even though the contents of the resource set won't change.

Use self.ranks instead of self.rlist to check for an empty resource
set and move the call out of the loop so it is made only once.

Fixes #5857