resource: round drain timestamps to nearest whole seconds
Problem: Timestamps for drained nodes are kept with sub-microsecond
precision. This causes the drain object in the resource.status response
to be much less compressible with arguably little benefit. It also
reduces the number of lines that can be coalesced in `flux resource
drain`.

Round timestamps while building the drainset object. This reduces
the drain object size when a significant number of ranks are drained
within the same second.