broker: avoid slow offline child UUID lookup
Problem: we're seeing the broker spend a lot of time in
child_lookup() on a system with a large, flat TBON when many
nodes are down.

When a message is received from a TBON child that is not currently
in the child hash because it is offline, it is either processed as
a new hello request, or sent a DISCONNECT control message.  Older
code used to log whether the child's UUID was known or unknown by
looking it up with a linear search in the child array.  The logging
has all been removed at this point, so the search can be dropped.

Fixes #5864