broker: reduce "online:" log messages
Problem: the broker "online: log messages are excessive on a system
instance.

2024-04-08T18:04:56.691537Z broker.info[0]: online: test[1,44-103,225-260,262-356,358-487,489-584,609-680
2024-04-08T18:04:56.879534Z broker.info[0]: online: test[1,44-103,225-260,262-356,358-487,489-584,609-680,705-728
etc

One problem is that when broker.quorum=1 (as it usually is for a system
instance), the 5s rate limit is bypassed.

But really the intent of these messages was to provide feedback during
wire-up, and that occurs during the QUORUM broker state.

Suppress the messages in all other broker states, which effectively
eliminates them for a system instance.