librlist: fix use-after-free after rlist_remove_ranks()
Problem: rlist_remove_rank() removes the requested rnode from the rlist
nodes list, but doesn't delete the item from the rank hash.  This can
lead to a use-after-free if the removed rank is looked up after this.

Call rank_hash_delete() after the rank is deleted from the nodes list
to ensure references to the removed rnode.