librlist: ignore missing ranks in rlist_mark_up/down
Problem: rlist_mark_state() doesn't check the return value from
rlist_find_rank(). If a rank is passed to rlist_mark_up() or
rlist_mark_down() that doesn't exist in the current rlist, then this
will result in a dereference of NULL and a segfault.

Skip ranks that don't exist in rlist_mark_start(), effectively
ignoring invalid ranks passed to rlist_mark_up() and rlist_mark_down().

This is simpler than returning an error, which would require multiple
iterations of the ranks idset as well as more complex error handling
on the caller side.