testsuite: check for working perms in t3203-instance-recovery.t
Problem: t3203-instance-recovery.t contains a couple tests that check
that an instance fails if the sqlite file is read or write-only.
However, this assumes that file permissions are actually working, and
in some environments (e.g. fakeroot) this is known not to be the case.

Check that file permissions are working in the test environment by
attempting to write to read-only file. Set a WORKING_PERMS prereq if
writing to the file fails and use this prereq on the affected tests.

Fixes #5862